+++ compose.yml
@@ -88,7 +88,6 @@ services:
       - node_modules_ghost_mw-vhost:/home/ghost/ghost/mw-vhost/node_modules:delegated
       - node_modules_ghost_nql-filter-expansions:/home/ghost/ghost/nql-filter-expansions/node_modules:delegated
       - node_modules_ghost_offers:/home/ghost/ghost/offers/node_modules:delegated
-      - node_modules_ghost_package-json:/home/ghost/ghost/package-json/node_modules:delegated
       - node_modules_ghost_payments:/home/ghost/ghost/payments/node_modules:delegated
       - node_modules_ghost_post-events:/home/ghost/ghost/post-events/node_modules:delegated
       - node_modules_ghost_post-revisions:/home/ghost/ghost/post-revisions/node_modules:delegated
@@ -253,7 +252,6 @@ volumes:
   node_modules_ghost_mw-vhost: {}
   node_modules_ghost_nql-filter-expansions: {}
   node_modules_ghost_offers: {}
-  node_modules_ghost_package-json: {}
   node_modules_ghost_payments: {}
   node_modules_ghost_post-events: {}
   node_modules_ghost_post-revisions: {}
