+++ ghost/core/core/server/lib/package-json/index.js
@@ -1,0 +1,1 @@
+module.exports = require('./package-json');
