+++ ghost/core/core/server/lib/package-json/package-json.js
