+++ ghost/core/core/server/lib/package-json/parse.js
