+++ ghost/core/package.json
@@ -136,7 +136,6 @@
     "@tryghost/mw-vhost": "0.0.0",
     "@tryghost/nodemailer": "0.3.45",
     "@tryghost/nql": "0.12.7",
-    "@tryghost/package-json": "0.0.0",
     "@tryghost/post-revisions": "0.0.0",
     "@tryghost/posts-service": "0.0.0",
     "@tryghost/pretty-cli": "1.2.44",
