--- ghost/package-json/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "@tryghost/package-json",
-  "private": true,
-  "version": "0.0.0",
-  "author": "Ghost Foundation",
-  "license": "MIT",
-  "main": "index.js",
-  "scripts": {
-    "dev": "echo \"Implement me!\"",
-    "test:unit": "NODE_ENV=testing c8 --all --reporter text --reporter cobertura --check-coverage -- mocha --reporter dot './test/**/*.test.js'",
-    "test": "yarn test:unit",
-    "lint:code": "eslint *.js lib/ --ext .js --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
-  },
-  "files": [
-    "index.js",
-    "lib"
-  ],
-  "devDependencies": {
-    "c8": "8.0.1",
-    "mocha": "10.2.0",
-    "should": "13.2.3",
-    "sinon": "15.2.0",
-    "tmp": "0.2.1"
-  },
-  "dependencies": {
-    "@tryghost/errors": "1.3.5",
-    "@tryghost/tpl": "0.1.32",
-    "fs-extra": "11.3.0",
-    "lodash": "4.17.21"
-  }
-}
